Devops/Docker/FLASK-POSTGRESQL/App/requirements.txt
flask 
psycopg2-binary